+++ src/components/@common/Card/Card.styles.ts
