+++ src/components/registerForm/CardCompany/CardCompany.stories.tsx
