+++ src/components/registerForm/CardCompany/CardCompany.styles.ts
