+++ src/components/registerForm/CardCompany/index.tsx
