+++ src/components/registerForm/CardPassword/CardPassword.styles.ts
