+++ src/components/registerForm/CardPassword/index.tsx
