+++ src/components/registerForm/SecurityCode/SecurityCode.styles.ts
