+++ src/pages/CardList/index.tsx
@@ -3,7 +3,7 @@ import Layout from "src/components/@common/Layout";
 import { Styled as S } from "./CardList.styles";
 import { useNavigate } from "react-router-dom";
 import useCardList from "src/hooks/useCardList";
-import Card from "src/components/@common/Temp";
+import Card from "src/components/@common/Card";
 import useModal from "src/hooks/useModal";
 import { PATHS } from "src/utils/constant";
